// client/src/prisma/models/moreitems.model.ts
/**
 * Copyright 2024 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/** MoreItems model represents the types of a set of more items */
export interface MoreItems {
  // index of list item
  index?: number;
  // icon of list item
  icon?: string;
  // text of the list item  */
  text?: string;
  // button type of the list item  */
  buttonType?: string;
  // changes the active state of list item
  activated?: boolean;
  // changes the visibility of list item
  visible?: boolean;
}